+++ crates/tiny/src/ui.rs
@@ -87,6 +87,10 @@ impl UI {
     delegate_ui!(set_nick(serv: &str, nick: &str,));
     delegate_ui!(set_tab_style(style: TabStyle, target: &MsgTarget,));
     delegate_ui!(user_tab_exists(serv_name: &str, nick: &str,) -> bool);
+
+    pub(crate) fn current_tab(&self) -> Option<MsgSource> {
+        self.ui.current_tab()
+    }
 }
 
 pub(crate) async fn task(