+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Logger now prints "Logging ended at ..." lines when you close a tab or exit
   tiny. (8061042)
 - Minor improvements in logging (d0505f2, bbb4b81)
+- `/join` (without arguments) now rejoins the current channel. (#334)
 
 # 2021/05/12: 0.9.0